Expose APIs for interacting with Azure credentials.
