+++ rush-plugins/rush-azure-storage-build-cache-plugin/src/AzureStorageBuildCacheProvider.ts
@@ -1,50 +1,24 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { ITerminal } from '@rushstack/node-core-library';
-import { PrintUtilities } from '@rushstack/terminal';
+import type { ITerminal } from '@rushstack/node-core-library';
 import {
   ICloudBuildCacheProvider,
-  CredentialCache,
-  ICredentialCacheEntry,
   EnvironmentVariableNames,
   RushConstants,
   EnvironmentConfiguration
 } from '@rushstack/rush-sdk';
+import { BlobClient, BlobServiceClient, BlockBlobClient, ContainerClient } from '@azure/storage-blob';
+
 import {
-  BlobClient,
-  BlobServiceClient,
-  BlockBlobClient,
-  ContainerClient,
-  ContainerSASPermissions,
-  generateBlobSASQueryParameters,
-  SASQueryParameters,
-  ServiceGetUserDelegationKeyResponse
-} from '@azure/storage-blob';
-import { DeviceCodeCredential, DeviceCodeInfo } from '@azure/identity';
+  AzureAuthorityHosts,
+  AzureStorageAuthentication,
+  IAzureStorageAuthenticationOptions
+} from './AzureStorageAuthentication';
 
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
-// TODO: This is a temporary workaround; it should be reverted when we upgrade to "@azure/identity" version 2.x
-// import { AzureAuthorityHosts } from '@azure/identity';
-export enum AzureAuthorityHosts {
-  AzureChina = 'https://login.chinacloudapi.cn',
-  AzureGermany = 'https://login.microsoftonline.de',
-  AzureGovernment = 'https://login.microsoftonline.us',
-  AzurePublicCloud = 'https://login.microsoftonline.com'
-}
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
-
-export type AzureEnvironmentNames = keyof typeof AzureAuthorityHosts;
-
-export interface IAzureStorageBuildCacheProviderOptions {
-  storageContainerName: string;
-  storageAccountName: string;
-  azureEnvironment?: AzureEnvironmentNames;
+export interface IAzureStorageBuildCacheProviderOptions extends IAzureStorageAuthenticationOptions {
   blobPrefix?: string;
-  isCacheWriteAllowed: boolean;
 }
-
-const SAS_TTL_MILLISECONDS: number = 7 * 24 * 60 * 60 * 1000; // Seven days
 
 interface IBlobError extends Error {
   statusCode: number;
@@ -57,14 +31,12 @@ interface IBlobError extends Error {
   };
 }
 
-export class AzureStorageBuildCacheProvider implements ICloudBuildCacheProvider {
-  private readonly _storageAccountName: string;
-  private readonly _storageContainerName: string;
-  private readonly _azureEnvironment: AzureEnvironmentNames;
+export class AzureStorageBuildCacheProvider
+  extends AzureStorageAuthentication
+  implements ICloudBuildCacheProvider
+{
   private readonly _blobPrefix: string | undefined;
   private readonly _environmentCredential: string | undefined;
-  private readonly _isCacheWriteAllowedByConfiguration: boolean;
-  private __credentialCacheId: string | undefined;
 
   public get isCacheWriteAllowed(): boolean {
     return EnvironmentConfiguration.buildCacheWriteAllowed ?? this._isCacheWriteAllowedByConfiguration;
@@ -73,12 +45,10 @@ export class AzureStorageBuildCacheProvider implements ICloudBuildCacheProvider 
   private _containerClient: ContainerClient | undefined;
 
   public constructor(options: IAzureStorageBuildCacheProviderOptions) {
-    this._storageAccountName = options.storageAccountName;
-    this._storageContainerName = options.storageContainerName;
-    this._azureEnvironment = options.azureEnvironment || 'AzurePublicCloud';
+    super(options);
+
     this._blobPrefix = options.blobPrefix;
     this._environmentCredential = EnvironmentConfiguration.buildCacheCredential;
-    this._isCacheWriteAllowedByConfiguration = options.isCacheWriteAllowed;
 
     if (!(this._azureEnvironment in AzureAuthorityHosts)) {
       throw new Error(
@@ -86,29 +56,6 @@ export class AzureStorageBuildCacheProvider implements ICloudBuildCacheProvider 
           `be one of: ${Object.keys(AzureAuthorityHosts).join(', ')}`
       );
     }
-  }
-
-  private get _credentialCacheId(): string {
-    if (!this.__credentialCacheId) {
-      const cacheIdParts: string[] = [
-        'azure-blob-storage',
-        this._azureEnvironment,
-        this._storageAccountName,
-        this._storageContainerName
-      ];
-
-      if (this._isCacheWriteAllowedByConfiguration) {
-        cacheIdParts.push('cacheWriteAllowed');
-      }
-
-      this.__credentialCacheId = cacheIdParts.join('|');
-    }
-
-    return this.__credentialCacheId;
-  }
-
-  private get _storageAccountUrl(): string {
-    return `https://${this._storageAccountName}.blob.core.windows.net/`;
   }
 
   public async tryGetCacheEntryBufferByIdAsync(
@@ -230,45 +177,6 @@ export class AzureStorageBuildCacheProvider implements ICloudBuildCacheProvider 
     }
   }
 
-  public async updateCachedCredentialAsync(terminal: ITerminal, credential: string): Promise<void> {
-    await CredentialCache.usingAsync(
-      {
-        supportEditing: true
-      },
-      async (credentialsCache: CredentialCache) => {
-        credentialsCache.setCacheEntry(this._credentialCacheId, credential);
-        await credentialsCache.saveIfModifiedAsync();
-      }
-    );
-  }
-
-  public async updateCachedCredentialInteractiveAsync(terminal: ITerminal): Promise<void> {
-    const sasQueryParameters: SASQueryParameters = await this._getSasQueryParametersAsync(terminal);
-    const sasString: string = sasQueryParameters.toString();
-
-    await CredentialCache.usingAsync(
-      {
-        supportEditing: true
-      },
-      async (credentialsCache: CredentialCache) => {
-        credentialsCache.setCacheEntry(this._credentialCacheId, sasString, sasQueryParameters.expiresOn);
-        await credentialsCache.saveIfModifiedAsync();
-      }
-    );
-  }
-
-  public async deleteCachedCredentialsAsync(terminal: ITerminal): Promise<void> {
-    await CredentialCache.usingAsync(
-      {
-        supportEditing: true
-      },
-      async (credentialsCache: CredentialCache) => {
-        credentialsCache.deleteCacheEntry(this._credentialCacheId);
-        await credentialsCache.saveIfModifiedAsync();
-      }
-    );
-  }
-
   private async _getBlobClientForCacheIdAsync(cacheId: string): Promise<BlobClient> {
     const client: ContainerClient = await this._getContainerClientAsync();
     const blobName: string = this._blobPrefix ? `${this._blobPrefix}/${cacheId}` : cacheId;
@@ -279,25 +187,7 @@ export class AzureStorageBuildCacheProvider implements ICloudBuildCacheProvider 
     if (!this._containerClient) {
       let sasString: string | undefined = this._environmentCredential;
       if (!sasString) {
-        let cacheEntry: ICredentialCacheEntry | undefined;
-        await CredentialCache.usingAsync(
-          {
-            supportEditing: false
-          },
-          (credentialsCache: CredentialCache) => {
-            cacheEntry = credentialsCache.tryGetCacheEntry(this._credentialCacheId);
-          }
-        );
-
-        const expirationTime: number | undefined = cacheEntry?.expires?.getTime();
-        if (expirationTime && expirationTime < Date.now()) {
-          throw new Error(
-            'Cached Azure Storage credentials have expired. ' +
-              `Update the credentials by running "rush ${RushConstants.updateCloudCredentialsCommandName}".`
-          );
-        } else {
-          sasString = cacheEntry?.credential;
-        }
+        sasString = await this.tryGetCachedCredentialAsync();
       }
 
       let blobServiceClient: BlobServiceClient;
@@ -320,51 +210,6 @@ export class AzureStorageBuildCacheProvider implements ICloudBuildCacheProvider 
     }
 
     return this._containerClient;
-  }
-
-  private async _getSasQueryParametersAsync(terminal: ITerminal): Promise<SASQueryParameters> {
-    const authorityHost: string | undefined = AzureAuthorityHosts[this._azureEnvironment];
-    if (!authorityHost) {
-      throw new Error(`Unexpected Azure environment: ${this._azureEnvironment}`);
-    }
-
-    const DeveloperSignOnClientId: string = '04b07795-8ddb-461a-bbee-02f9e1bf7b46';
-    const deviceCodeCredential: DeviceCodeCredential = new DeviceCodeCredential(
-      'organizations',
-      DeveloperSignOnClientId,
-      (deviceCodeInfo: DeviceCodeInfo) => {
-        PrintUtilities.printMessageInBox(deviceCodeInfo.message, terminal);
-      },
-      { authorityHost: authorityHost }
-    );
-    const blobServiceClient: BlobServiceClient = new BlobServiceClient(
-      this._storageAccountUrl,
-      deviceCodeCredential
-    );
-
-    const startsOn: Date = new Date();
-    const expires: Date = new Date(Date.now() + SAS_TTL_MILLISECONDS);
-    const key: ServiceGetUserDelegationKeyResponse = await blobServiceClient.getUserDelegationKey(
-      startsOn,
-      expires
-    );
-
-    const containerSasPermissions: ContainerSASPermissions = new ContainerSASPermissions();
-    containerSasPermissions.read = true;
-    containerSasPermissions.write = this._isCacheWriteAllowedByConfiguration;
-
-    const queryParameters: SASQueryParameters = generateBlobSASQueryParameters(
-      {
-        startsOn: startsOn,
-        expiresOn: expires,
-        permissions: containerSasPermissions,
-        containerName: this._storageContainerName
-      },
-      key,
-      this._storageAccountName
-    );
-
-    return queryParameters;
   }
 
   private _getConnectionString(sasString: string | undefined): string {
