+++ rush-plugins/rush-azure-storage-build-cache-plugin/src/index.ts
@@ -2,5 +2,11 @@
 // See LICENSE in the project root for license information.
 
 import { RushAzureStorageBuildCachePlugin } from './RushAzureStorageBuildCachePlugin';
+export {
+  AzureStorageAuthentication,
+  IAzureStorageAuthenticationOptions,
+  AzureEnvironmentNames,
+  AzureAuthorityHosts
+} from './AzureStorageAuthentication';
 
 export default RushAzureStorageBuildCachePlugin;
